Rename Ncco interface and inheritors to Action as the whole JSON array is a single NCCO.
