+++ src/main/java/com/nexmo/client/voice/ncco/ConversationAction.java
@@ -27,68 +27,76 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonInclude(value = JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class RecordNcco implements Ncco {
-    private static final String ACTION = "record";
+public class ConversationAction implements Action {
+    private static final String ACTION = "conversation";
 
-    private RecordingFormat format = null;
-    private Integer endOnSilence = null;
-    private Character endOnKey = null;
-    private Integer timeout = null;
-    private Boolean beepStart = null;
+    private String name;
+    private String musicOnHoldUrl = null;
+    private Boolean startOnEnter = null;
+    private Boolean endOnExit = null;
+    private Boolean record = null;
     private String[] eventUrl = null;
     private String eventMethod = null;
-    private SplitRecording split = null;
 
-    public RecordingFormat getFormat() {
-        return format;
+
+    public ConversationAction(@JsonProperty("name") String name) {
+        this.name = name;
     }
 
-    public void setFormat(RecordingFormat format) {
-        this.format = format;
+    public String getName() {
+        return name;
     }
 
-    public Integer getEndOnSilence() {
-        return endOnSilence;
+    public void setName(String name) {
+        this.name = name;
     }
 
-    public void setEndOnSilence(Integer endOnSilence) {
-        this.endOnSilence = endOnSilence;
+    public String getMusicOnHoldUrl() {
+        return musicOnHoldUrl;
     }
 
-    public Character getEndOnKey() {
-        return endOnKey;
+    @JsonProperty("musicOnHoldUrl")
+    public String[] getMusicOnHoldUrlAsArray() {
+        // TODO: Rework in 4.0.
+        // This property is expected to be serialized as an array, however we want to also insure it remains null
+        // if null.
+        return this.musicOnHoldUrl != null ? new String[]{this.musicOnHoldUrl} : null;
     }
 
-    public void setEndOnKey(Character endOnKey) {
-        this.endOnKey = endOnKey;
+    public void setMusicOnHoldUrl(String... musicOnHoldUrl) {
+        this.musicOnHoldUrl = musicOnHoldUrl[0];
     }
 
-    public Integer getTimeout() {
-        return timeout;
+    public Boolean getStartOnEnter() {
+        return startOnEnter;
     }
 
-    public void setTimeout(Integer timeout) {
-        this.timeout = timeout;
+    public void setStartOnEnter(Boolean startOnEnter) {
+        this.startOnEnter = startOnEnter;
     }
 
-    public Boolean getBeepStart() {
-        return beepStart;
+    public Boolean getEndOnExit() {
+        return endOnExit;
     }
 
-    public void setBeepStart(Boolean beepStart) {
-        this.beepStart = beepStart;
+    public void setEndOnExit(Boolean endOnExit) {
+        this.endOnExit = endOnExit;
+    }
+
+    public Boolean getRecord() {
+        return record;
+    }
+
+    public void setRecord(Boolean record) {
+        this.record = record;
     }
 
     public String[] getEventUrl() {
         return eventUrl;
     }
 
-    public void setEventUrl(String eventUrl) {
-        setEventUrl(new String[]{eventUrl});
-    }
-
     @JsonProperty("eventUrl")
-    public void setEventUrl(String[] eventUrl) {
+    public void setEventUrl(String... eventUrl) {
         this.eventUrl = eventUrl;
     }
 
@@ -103,14 +111,6 @@ public class RecordNcco implements Ncco {
     @Override
     public String getAction() {
         return ACTION;
-    }
-
-    public SplitRecording getSplit() {
-        return split;
-    }
-
-    public void setSplit(SplitRecording split) {
-        this.split = split;
     }
 
     @Override
