+++ src/main/java/com/nexmo/client/voice/ncco/StreamAction.java
@@ -24,29 +24,35 @@ package com.nexmo.client.voice.ncco;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.nexmo.client.voice.VoiceName;
 
 @JsonInclude(value = JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class TalkNcco implements Ncco {
-    private static final String ACTION = "talk";
+public class StreamAction implements Action {
+    private static final String ACTION = "stream";
 
-    private String text;
+    private String streamUrl;
+    private Float level = null;
     private Boolean bargeIn = null;
     private Integer loop = null;
-    private Float level = null;
-    private VoiceName voiceName = null;
 
-    public TalkNcco(@JsonProperty("text") String text) {
-        this.text = text;
+    public StreamAction(@JsonProperty("streamUrl") String streamUrl) {
+        this.streamUrl = streamUrl;
     }
 
-    public String getText() {
-        return text;
+    public String getStreamUrl() {
+        return streamUrl;
     }
 
-    public void setText(String text) {
-        this.text = text;
+    public void setStreamUrl(String streamUrl) {
+        this.streamUrl = streamUrl;
+    }
+
+    public Float getLevel() {
+        return level;
+    }
+
+    public void setLevel(Float level) {
+        this.level = level;
     }
 
     public Boolean getBargeIn() {
@@ -63,24 +69,6 @@ public class TalkNcco implements Ncco {
 
     public void setLoop(Integer loop) {
         this.loop = loop;
-    }
-
-    public void setLevel(Float level) {
-        this.level = level;
-    }
-
-    public Float getLevel() {
-        return level;
-    }
-
-    // TODO: Change to return VoiceName on next major version
-    public String getVoiceName() {
-        return (voiceName == null) ? null : voiceName.toString();
-    }
-
-    // TODO: Change to accept VoiceName on next major version
-    public void setVoiceName(String voiceName) {
-        this.voiceName = VoiceName.fromString(voiceName);
     }
 
     @Override
