+++ src/test/java/com/nexmo/client/voice/ncco/TalkActionTest.java
@@ -22,23 +22,22 @@
 package com.nexmo.client.voice.ncco;
 
 import com.nexmo.client.voice.VoiceName;
-
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 
-public class TalkNccoTest {
+public class TalkActionTest {
     @Test
     public void testJsonHasTalkAndActionWhenGivenOnlyRequiredParameters() {
         String json = "{\"text\":\"Talk to me\",\"action\":\"talk\"}";
-        TalkNcco ncco = new TalkNcco("Talk to me");
+        TalkAction ncco = new TalkAction("Talk to me");
         assertEquals(json, ncco.toJson());
     }
 
     @Test
     public void testJsonHasChangedTalkWhenChanged() {
         String json = "{\"text\":\"Still talk to me\",\"action\":\"talk\"}";
-        TalkNcco ncco = new TalkNcco("Talk to me");
+        TalkAction ncco = new TalkAction("Talk to me");
         ncco.setText("Still talk to me");
         assertEquals(json, ncco.toJson());
     }
@@ -46,7 +45,7 @@ public class TalkNccoTest {
     @Test
     public void testJsonHasBargeInWhenProvidedAndIsTrue() {
         String json = "{\"text\":\"Talk to me\",\"bargeIn\":true,\"action\":\"talk\"}";
-        TalkNcco ncco = new TalkNcco("Talk to me");
+        TalkAction ncco = new TalkAction("Talk to me");
         ncco.setBargeIn(true);
         assertEquals(json, ncco.toJson());
     }
@@ -54,7 +53,7 @@ public class TalkNccoTest {
     @Test
     public void testJsonHasBargeInWhenProvidedAndIsFalse() {
         String json = "{\"text\":\"Talk to me\",\"bargeIn\":false,\"action\":\"talk\"}";
-        TalkNcco ncco = new TalkNcco("Talk to me");
+        TalkAction ncco = new TalkAction("Talk to me");
         ncco.setBargeIn(false);
         assertEquals(json, ncco.toJson());
     }
@@ -62,7 +61,7 @@ public class TalkNccoTest {
     @Test
     public void testJsonHasLoopWhenProvided() {
         String json = "{\"text\":\"Talk to me\",\"loop\":3,\"action\":\"talk\"}";
-        TalkNcco ncco = new TalkNcco("Talk to me");
+        TalkAction ncco = new TalkAction("Talk to me");
         ncco.setLoop(3);
         assertEquals(json, ncco.toJson());
     }
@@ -70,7 +69,7 @@ public class TalkNccoTest {
     @Test
     public void testJsonHasLevelWhenProvided() {
         String json = "{\"text\":\"Talk to me\",\"level\":-0.34,\"action\":\"talk\"}";
-        TalkNcco ncco = new TalkNcco("Talk to me");
+        TalkAction ncco = new TalkAction("Talk to me");
         ncco.setLevel(-0.34f);
         assertEquals(json, ncco.toJson());
     }
@@ -78,7 +77,7 @@ public class TalkNccoTest {
     @Test
     public void testJsonHasVoiceNameWhenProvided() {
         String json = "{\"text\":\"Talk to me\",\"voiceName\":\"Kimberly\",\"action\":\"talk\"}";
-        TalkNcco ncco = new TalkNcco("Talk to me");
+        TalkAction ncco = new TalkAction("Talk to me");
         ncco.setVoiceName(VoiceName.KIMBERLY.toString());
         assertEquals(json, ncco.toJson());
     }
